+++ mybooking-templates/mybooking-plugin-activities.php
@@ -19,12 +19,21 @@
 		<!-- Products -->
 		<div class="row">
 			<?php foreach( $args['data']->data as $mybooking_activity ) { ?>
+			  <?php  $mybooking_activityIdAnchor = $mybooking_activity->id;
+	  	    if ( !empty( $mybooking_activity->slug) ) {
+	  			  $mybooking_activityIdAnchor = $mybooking_activity->slug;
+  				}
+			  ?>
 			  <div class="col-lg-4">
 					<div class="activity-card card mb-2">
 						<?php if ( !empty( $mybooking_activity->photo_url_full ) ) { ?>
-								  <img class="activity-card-img card-img-top" src="<?php echo esc_url( $mybooking_activity->photo_url_full )?>" alt="<?php echo esc_attr( $mybooking_activity->name )?>">
+						  <a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_activityIdAnchor ) ?>">
+						  	<img class="activity-card-img card-img-top" src="<?php echo esc_url( $mybooking_activity->photo_url_full )?>" alt="<?php echo esc_attr( $mybooking_activity->name )?>">
+						  </a>	
 						<?php } else { ?>
-							   <img class="activity-card-img card-img-top" src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/default-image-product.png' ) ?>" alt="<?php echo esc_attr( $mybooking_activity->name )?>">
+							<a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_activityIdAnchor ) ?>">
+							  <img class="activity-card-img card-img-top" src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/default-image-product.png' ) ?>" alt="<?php echo esc_attr( $mybooking_activity->name )?>">
+							</a>   
 						<?php } ?>
 					  <div class="card-body d-flex flex-column justify-content-center">
 					    <h3 class="h6 card-title activity-card-title"><?php echo esc_html( $mybooking_activity->name ) ?></h3>
@@ -40,12 +49,7 @@
 					  	<?php } ?>
 					  </div>
 					  <div class="card-body activity-more-information d-flex flex-column justify-content-center">
-					  	<?php  $mybooking_activityIdAnchor = $mybooking_activity->id;
-					  	    if ( !empty( $mybooking_activity->slug) ) {
-					  			  $mybooking_activityIdAnchor = $mybooking_activity->slug;
-  			  				}
-					  		?>
-					    	<a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_activityIdAnchor ) ?>" class="btn btn-primary"><?php echo esc_html_x( 'More information', 'activities_list', 'mybooking-wp-plugin' ) ?></a>
+				    	<a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_activityIdAnchor ) ?>" class="btn btn-primary"><?php echo esc_html_x( 'More information', 'activities_list', 'mybooking-wp-plugin' ) ?></a>
 					  </div>
 					</div>
 			  </div>