+++ mybooking-templates/mybooking-plugin-products.php
@@ -10,12 +10,23 @@
 		<!-- Products -->
 		<div class="row">
 			<?php foreach( $args['data']->data as $mybooking_product ) { ?>
+
+			  <?php  $mybooking_productIdAnchor = $mybooking_product->id;
+	  	    if ( !empty( $mybooking_product->slug) ) {
+	  			  $mybooking_productIdAnchor = $mybooking_product->slug;
+  				}
+			  ?>
+
 			  <div class="col-md-4">
 					<div class="resource-product-card card d-flex flex-column mb-2 shadow mb-5 bg-white rounded-0">
 						<?php if ( !empty( $mybooking_product->photo_path ) ) { ?>
-					  	<img class="resource-product-card-img card-img-top rounded-0" src="<?php echo esc_url( $mybooking_product->photo_path )?>" alt="<?php echo esc_attr( $mybooking_product->name )?>">
+							<a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_productIdAnchor ) ?>">
+					  	  <img class="resource-product-card-img card-img-top rounded-0" src="<?php echo esc_url( $mybooking_product->photo_path )?>" alt="<?php echo esc_attr( $mybooking_product->name )?>">
+					    </a>
 						<?php } else { ?>
- 					    <img class="resource-product-card-img card-img-top rounded-0" src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/default-image-product.png' ) ?>" alt="<?php echo esc_attr( $mybooking_product->name )?>">							
+							<a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_productIdAnchor ) ?>">
+	 					    <img class="resource-product-card-img card-img-top rounded-0" src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/default-image-product.png' ) ?>" alt="<?php echo esc_attr( $mybooking_product->name )?>">
+	 					  </a>  							
 						<?php } ?>
 					  <div class="card-body d-flex flex-column justify-content-center">
 					    <h5 class="h6 card-title text-left resource-product-card-title"><b><?php echo esc_html( $mybooking_product->name ) ?></b></h5>
@@ -35,7 +46,8 @@
                 <?php if ( isset( $mybooking_product->characteristic_length ) && !empty( $mybooking_product->characteristic_length ) ) { ?>
                   <li class="icon-list-item">
                       <span class="icon-list-icon">
-                      	<img src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/characteristics/length.svg' ) ?>"/>
+                      	<img src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/characteristics/length.svg' ) ?>"
+                      	     alt="<?php echo esc_attr( MyBookingEngineContext::getInstance()->getCharacteristic( 'length' ) ) ?>"/>
                       </span>
                       <span class="icon-list-text text-muted"><?php echo wp_kses( sprintf( _x('<b>%.2f</b> m.', 'renting_products', 'mybooking-wp-plugin' ), 
                       	                                                                   $mybooking_product->characteristic_length ),
@@ -45,7 +57,8 @@
 								<?php foreach ( $mybooking_product->key_characteristics as $mybooking_key => $mybooking_value) { ?>
 									<li class="icon-list-item">
 											<span class="icon-list-icon">
-												<img src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/key_characteristics/'.$mybooking_key.'.svg' ) ?>"/>
+												<img src="<?php echo esc_url( plugin_dir_url(__DIR__).'/assets/images/key_characteristics/'.$mybooking_key.'.svg' ) ?>"
+												     alt="<?php echo esc_attr( MyBookingEngineContext::getInstance()->getKeyCharacteristic( $mybooking_key ) ) ?>"/>
 											</span>
 											<span class="icon-list-text text-muted"><?php echo esc_html( $mybooking_value ) ?></span>
 									</li>
@@ -54,7 +67,7 @@
 						  </div>	
 			      <?php } ?>
 					  <div class="card-body d-flex flex-column justify-content-end">
-					    <a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_product->code ) ?>" class="btn btn-primary w-100"><?php echo esc_html_x('More information', 'renting_products', 'mybooking-wp-plugin' ) ?></a>
+					    <a href="<?php echo esc_url( '/'.$args['url_detail'].'/'.$mybooking_productIdAnchor ) ?>" class="btn btn-primary w-100"><?php echo esc_html_x('More information', 'renting_products', 'mybooking-wp-plugin' ) ?></a>
 					  </div>
 					</div>
 			  </div>